SignConnect/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.123.5
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0

numpy
tensorflow
cv2
mediapipe
sklearn.model_selection
sklearn.preprocessing
tensorflow.keras.model
tensorflow.keras.layers